requirements.txt
pandas
jupyterlab
matplotlib.pyplot
seaborn
tabulate